FEAT: connection struct
