+++ server/Cargo.toml
@@ -13,3 +13,4 @@ env_logger = "0.10.1"
 webrtc = "0.9.0"
 actix-session = { version = "0.9.0", features = ["cookie-session"] }
 rand = "0.8.5"
+bytes = "1.5.0"
